Remove unnecessary operations module variables
The idea was to move away from creating variables for operational types
that do not actually allow for decisions to be made about power output,
e.g. must-run and variable generators for now. Instead, there is now an
expression 'Power_Provision' that is equal to the variable
Provide_Power in the case of dispatchable generators, but to a
parameter or an arithmetic expression including several parameters in
the case of must-run and variable generators respectively. Since no
variables are created for them, the must-run and variable modules also
do not require that any constraints on those variables be created.