Determine needed operational modules from inputs
Instead of having to manually specify the operational modules needed,
dynamically determine which operational modules will have to be loaded
based on the input data for a scenario. This is accomplished by
reading the 'operational_type' column of generators.tab and finding
the unique set of operational types. An operational module is loaded
only if a generator of this operational type exists in the input data.
For example, if no variable generators exist in the data, the
'variable' operational module won't be loaded; any module-specific
attributes will therefore also not be called, thus preventing the need
to handle erros resulting from, say, not finding an input file
associated with that module or similar issues.