Implement startup and shutdown costs
The operations.py module, in its determine_dynamic_components function,
now decides which generators have startup and/or shutdown cost that
shoud be added to the objective function. For each generator in
the generators.tab input file, if the startup_cost column contains a
numerical value greater than 0, the generator is added to the
m.startup_cost_generator list, which is then used to initialize the
STARTUP_COST_GENERATORS set of the abstract model. Similary, if the
shutdown_cost column contains a numerical value greater than 0,
the generator is added to the m.shutdown_cost_generator list, which is
then used to initialize the SHUTDOWN_COST_GENERATORS set of the
abstract model. When a numerical value greater than 0 is found for
either startup or shutdown cost in generators.tab, the value is also
added to the respective dictionary (m.startup_cost_generators or
m.shutdown_cost_generators) with the generator as a key; these
dictionaries are used to initialize the startup_cost and shutdown_cost
param components of the abstract model.

In each operational module, rules are added for what a 'startup' and a
'shutdown' is that are used as to define the Startup_Expression and
Shutdown_Expression components of the model in operations.py.

Two additional variables are defined:

Startup_Cost, indexed over STARTUP_COST_GENERATORS and TIMEPOINTS
Shutdown_Cost, indexed over SHUTDOWN_COST_GENERATORS and TIMEPOINTS

Each of these variables has non-negative real numbers as its domain.
These variables are constrained to be greater than or equal to
Startup_Expression * startup_cost and
Shutdown_Expression * shutdown_cost respectively.

Startup_Expression (when units were actually shut down since the
previous timepoint) and Shutdown_Expression (when units were actually
started up since the previous timepoint) may be negative, but the
non-negative domain of Startup_Cost and Shutdown_Cost means that, in a
cost minimization, their values will be pushed to 0, but can't go
negative.

For now, the 'previous timepoint' for the first timepoint in the data
is defined as the last timepoint in the day -- a circular treatment of
the dispatch horizon.