Aggregate costs in objective function module
Moved the summing up of individual generator costs to the objective
function module (objective.min_total_cost). The goal is to keep all
objective function elements in one place, as it's cleaner and easier to
debug. The implementation remains modular -- empty expressions will
be passed to the objective if a module has not been called or a model
component has been initialized with an empty dataset.