Export operations results to CSV
Results from the operations module are now exported to a CSV in the
scenario 'directory' named operations.csv. This is accomplished by
creating pandas dataframes for each variable or, in the case of power
provision, an expression, indexed by generator (the subset of
generators that have that variable) and timepoint. Individual modules
can add their own dataframes. The dataframes are then merged and
exported to a CSV.

Also:
Move functions used to manipulate data in operations.py to their own
module, called auxiliary.py.
Some clean-up of imports of modules.
Committing objective_function_value.txt and Meet_Load_Constraint.csv
in results.