Add 'horizons' functionality
Introduce the concept of 'horizons.' Horizons are combinations of and
relationships among timeponts. For example, a day can be a
horizon containing 24 timepoints (hourly operations) or 288 timepoints
(5-minute operations) or a week could be a horizon containing 7
timepoints (daily operations). 'Horizons' link subsets of timepoints
together. Constraints can then be imposed for each horizon: an example
would be a daily energy budget constraint for hydro or an energy
balance constraint for storage.

The boundaries of the horizon needs to be defined and two options are
implemented here: 'circular' boundary and 'linear' boundary. In the
former, the last timepoint of the horizon is defined as the previous
timepoint of the first timepoint of the horizon; in the latter, no
previous timepoint is defined for the first timepoint of the horizon
and therefore any variable, expression, and constraints must be
excluded in the 'linear' boundary case. Tracking the previous timepoint
for each timepoint is currenty needed for the startup and shutdown
functionality and will be needed for tracking storage state of charge
and maybe other functionality in the future. The boundary type can be
defined for each horizon individually.