Pass final generator commitment to later stages
First working implementation of functionality to pass final generator
commitment to the following stages in multi-stage scenarios. This is
accomplished as follows:

A new column is needed in generators.tab that indicates the name of the
final commitment stage for each generator (that has a commitment
variable).

Changes to run_scenario.py including:
1. in the ScenarioStructure class, add functionality to create a file
named fixed_commitment.csv in subdirectory named 'pass_through_inputs'
if a subproblems file is found in each horizon's directory; add header
to this file; each stage will write the values of commitment variables
that are to be fixed in the following stages to this file; the file is
replaced with a new file any time the scenario is run, so old results
will be deleted upon re-running
2. fix_variables function that calls the fix_variables function, if it
exists, of each loaded module
3. call the fix_variables function after the instance is created to
make the changes (variable fixes) to the instance; if no modules call
for fixing of variables, this does nothing to the original instance

The dispatchable_binary_commit and dispatchable_continuous_commit
modules contain two new functions:
1. commitment_rule: returns the respective commitment variable for that
type of generators
2. fix_commitment: sets commitment variable to a pre-specified value,
fixed_commitment, which is determined by the previous stage, and fixes
the variable

A new module, fix_commitment under generation.operations does the
following:
<determine_dynamic_components>
1. reads the final commitment stage for each generator and compares
that to the current stage to create the set of generators for which
the current stage is their final commitment stage; the current stage
is currently based on the problem directory name
2. reads the fixed commitment of any generators whose last commitment
stage was before the current stage; the fixed commitment is in a CSV
file located in a directory named pass_through_inputs in each horizon's
directory; when the current stage run begins, all prior stages have
already written to this file any generator commitment that is to be
fixed in the current stage
<add_model_components>
1. creates set of generators for which the current stage is the final
commitment stage and an expression with the respective commitment
variable for every generator, which is created by calling the
commitment_rule from the generator's operational module (the expression
is needed in the export function)
2. creates set of generators whose commitment has already been fixed
and the parameter with their commitment
<fix_variables>
1. loads the operational modules
2. calls the fix_commitment function from the operational module of
each generator whose commitment is to be fixed in the current stage
<export_results>
1. opens the fixed_commitment.csv file in the pass_through_inputs
directory
2. loops through the generators fow which the current stage is their
final commitment stage and appends the generator name, timepoint,
generator's final commitment stage, and the commitment variable value
to the file