Common-input directories for horizons and stages
Since horizon subproblems and stage subproblems share some data inputs,
these data are now in a subdirectory named 'inputs' in the root
directory the horizons or stages share. The inputs that vary by horizon
and/or stage are still in a directory named 'inputs' in each horizon
and/or stages's directory.

The horizon and stage names are now passed to the modules. Each module
is responsible for finding its own input data (i.e. whether or not to
use the horizon and stages names in determining the input file path).