Separate class for dynamically determined inputs
Create new class and set dynamically determined data inputs as
attributes to that new class rather than to the AbstractModel class,
which was the previous approach. These inputs are used to initialize
model components such as sets and params upon adding them to the model
and to keep track of which variables to assign to which set members,
etc. They are not model components per se and there is enough data at
this point that keeping those inputs as attributes to the model could
get problematic. Instead, having a separate data object is more
appropriate and can also be individually controlled would allow greater
flexibility to, say, free up memory, if needed.