Initialize dynamic data components with BuildAction
Some parameters are specified for only some generatos, e.g. startup
costs or fuel (as opposed to parameters such as lf_reserves_up, which
must be specified for all generators). The goal is to load such
parameters ONLY for the generators they are specified for without
having to rely on separate input files or on fake values for the rest
of the generators. This was previously accomplished by parsing the
input file in the determine_dynamic_inputs function of a given module
and the using initialize to construct the data-dependent model sets and
params. One problem was that this required keeping a separate data
object around for use when components were initialized, increasing
memory requirements if not deleted, etc.

Instead of initializing these model components from a separate data
object and before the rest of the data are loaded, the new treatment
uses the BuildAction function of Pyomo to construct them instead.
BuildAction is a callback function that makes possible the construction
of components when the model is created. The BuildAction object will
call a function that can initialize a model component from a Python
data structure. Adding all the abstract model components first, loading
the REQUIRED data next via a DataPortal next, and finally determining
the DYNAMIC inputs and initializing the associated model components
with BuildAction when the model is created is a more intuitive workflow.

The determine_dynamic_inputs function of modules has been renamed to
determine_dynamic_components to better reflect that it does not deal
with loading data anymore. Rather, the purpose of the
determine_dynamic_components function is now to keep track of which, if
any, dynamic model components are associated with each index.