Initial implementation of (investment) periods
This is the first attempt at implementing 'periods,' an index over
which generator capacities can change and that will eventually allow
for endogenous 'build' variables, 'retirement' variables, etc.
The approach will be similar to that take with the operations module:
a 'capacity_type' will be defined for each generator (analogous to the
'operational_type'). The capacity type will determine the particular
params, variables, expressions, and constraints that will be relevant
to each generator. Currently, a single capacity type is implemented --
specified_no_economic_retirement -- which is the type for generators
with exogenously specified capacities and no retirement allowed.
Eventually, other capacity types may include
'specified_linear_economic_retirement,'
'specified_binary_economic_retirement,' 'new_build,' etc.
The capacity_type modules will pass information to the main capacity
module (modules.generation.capacity.capacity). Currently, the only
information passed is for the Capacity_MW expression for each
generator and period.

With multiple periods across which generator capacities can change
(i.e. generators can be built or retired), it also becomes necessary
to keep track of when generators are available to be operated. Each
capacity_type will have a two-dimensional set of the generator-period
combinations when a generator is operational (the set will be
initialized by the respective capacity_type module). These sets are
used to create the GENERATOR_OPERATIONAL_TIMEPOINTS set, a two-
dimensional set of generator-timepoint combinations when a generator is
operational. This is the set over which generator operational
constraints can be imposed to avoid creating constraints for
generator-timepoint combinations when a generator is not actually
available. Subsets of this set based on operational type are also used.