Initial implementation of 'new builds'
Introduce a new capacity_type -- new_build_option -- which allows the
model to build new capacity with a certain lifetime for a certain cost.
The lifetime (the parameter lifetime_yrs_by_new_build_vintage)
determines how long the capacity build in a certain period (i.e.
capacity of a certain vintage) will be operational. This is
accomplished by creating the NEW_BUILD_OPTION_OPERATIONAL_PERIODS set,
which is then joined with the equivalent set for the
specified_no_economic_retirements capacity type, i.e.
SPECIFIED_NO_ECON_RETRMNT_GENERATORS_OPERATIONAL_PERIODS, to create the
GENERATOR_OPERATIONAL_PERIODS superset. The strategy is for capacity
modules, if called, to add their equivalent sets to be joined to create
the superset by adding their set name to the newly added
capacity_type_operational_period_sets list attribute of the abstract
model class. This GENERATOR_OPERATIONAL_PERIODS set is used to create
sets of timepoints over which generators are operational that are in
turn used to index variables and constraints in the operational module.
The goal is to only create variables and constraints for generators
that are actually available to be operated rather than carry zeros
around for non-operational generators' capacities and create
unnecessary model components.

The cost (the parameter annualized_real_cost_per_mw_yr) is applied
in each operational period and added to the objective function. These
costs are multiplied by two parameters indexed by period:
discount_factor (costs further in the future may be weighted less than
costs closer to the present) and number_years_represented (each period
can be weighted more or less depending on the length of time it
represents). All operational costs are now also multiplied by these two
parameters.