Add transmission capacity types
Add transmission capacity types in preparation for implementing the
ability to build new transmission. The treatment is similar to that of
generation and storage. The main transmission module (transmission.py
in the geography module) calls on functions from the transmission
capacity type modules to determine transmission capacity, which then
is used to constrain transmission operations.

Also fixed bug -- Transmit_Power was not constrained before.