Initial implementation of RPS targets
The current implementation allows for the RPS zones to be different
from the load zones. Each generator that can contribute to the target
of an RPS zone must have that rps_zone specified in generators.tab.
For RPS purposes also implemented a curtailment variable for
generators of the 'variable' operational type. The power provision
from those generators is now the capacity times capacity factor minus
curtailment. That is what gets added to the load balance and to the
RPS constraint. In this treatment, curtailment is by generator, i.e.
variable generators are in a way treated like 'dispatchable' generators
-- they can be 'dispatched' downward.