Split project-level reserves to own modules
These are grouped with the system-level reserve modules, so that if
a user specifies lf_reserves_up for example, both the project-level
and system-level lf_reserves_up modules will be included (otherwise,
they will be deleted from the list of requested modules).

The dynamic component creation of the generation_and_storage module is
now in its __init__ file -- that way, the resources.tab file is only
read in once. All of the components read in the __init__ file are
required regardless of which optional modules are specified. Putting
this in __init__ allows us to have to read resources.tab only once.