Standardize system-level reserves module
All reserves submodules now call generic functions from the standard
module.

A few differences in objective functions are due to the fact that
loading the reserve violations penalty from a file is slightly
different from initializing the param in the model (ends in 0 instead
of 9). Problem files are the same other than that.