Explicitly link modules to dynamic component names
Different modules can use the same attribute from the DynamicComponents
class instance; by 1) importing the name of the attribute from the
dynamic_components module into each module that modifies or uses that
attribute and 2) using the name as an argument to a getattr() or
setattr() function, we can more explicitly link the modules, track
where components are created, and facilitate refactoring (e.g. easier
to use IDE built-in rename refactor instead of having to do recursive
searches for the component name).