Create min and max power constraints in op modules
Previously these constraints were created in operations.py for all
projects with the appropriate function being called from the
operational module (which would return Constraint.Skip for some
projects). It's cleaner to just have all the operational constraints
for a given operational type in one place, plus there is no reason to
aggregate for min and max power in the same way as, say, for power
provision.

The objective function for caiso_initial_test changed slightly. None
of the other example runs had a change in the objective.