Fix bug in cumulative cap constraint data-loading
Pandas read_csv takes the usecols param through which one can specify
which columns from an input file to read. These columns must be a
subset of the columns actually in the file. If the columns do not
exist, this will throw an error -- except if the number of columns
specified via usecols is the same as the number of columns in the input
file. This was therefore working for a weird reason before: some
semi-buggy pandas behavior.