Subhourly energy adjustment for variable op type
Add functionality to adjust curtailment and delivered energy for
variable generators when they provide reserves. This is particularly
relevant for RPS-eligible generators for which we have to track how
much of the available energy was delivered and how much was curtailed.
When providing downward reserves, these generators incur additional
curtailment within the hour when called upon (on top of any curtailment
scheduled for the hour/timepoint). As not all the scheduled power is
delivered in such cases, we must subtract this subhourly curtailment
from the delivered RPS energy. Conversely, when providing upward
reserves, these generators will occasionally be called upon to increase
ther power production -- this additional delivered energy must be added
to delivered RPS energy.

The subhourly energy adjustment is implemented as a simple linear
function of reserve provision. It can be different for each type of
reserve and can also vary by balancing area within each reserve type.
Data are loaded through the reserve_type_balancing_areas.tab file.

There are some changes to the headers of these files to simplify and
standardize data input.

There are also a couple of additional test cases as well as small
additions to the tests for the project/operations/reserves modules.