Fix underconstrained generic storage
Generic storage resources were underconstrained. Additional or
reformulated constraints for storage now include:
1. Constrain charging and dicharging to less than or equal to the
project power capacity
2. Constrain headroom (sum of all upward reserves) to less than or
equal to [(capacity_mw - dicharging_mw) + charging_mw]
3. Constrain footroom (sum of all downward reserves) to less than or
equal to [(capacity_mw - charging_mw) + discharging_mw]
4. Constrain headroom to the available energy in storage (see caveats
and TODOs)
5. Constrain footroom to the available storage (i.e. 'room in the tank')
(see caveats and TODOs)

For 4 and 5, the assumption is that in order to provide reserves, the
project would have to be able to sustain the new setpoint for 1 hour.
This should become a parameter that can be varied.

Also for 4 and 5, the assumption is that the net energy gained/lost
during the current timepoint counts in the constraints. In reality,
reserves could be called at the very beginning or the very end of the
timepoint (e.g. hour). If called at the end, we would have all the net
energy (or resulting 'room in the tank') available, but if called in
the beginning none of it would be available. The net energy should be
derated in those constraints and should be a param that can be varied.

Also fixed a bug in the energy tracking constraint where the charging
and discharging variables for the previous timepoint were incorrectly
being  multiplied by the 'hours in timepoint' param for the current
timepoint.

Updated the objective function values for several of the tests. Storage
in those examples in now more constrained, so the objective function
values increase.