Refactor project operations package
The project.operations.operations module is now named
project.operations power. It only includes that Power_Provision_MW
expression for all projects and the aggregation of power provision to
load_zone-timepoint.

The curtailment expressions are now in project.operations.curtailment,
a module that is only currently needed if the RPS supermodule is
requested by the user.

The calling of functions from the operational type modules is now in
the __init__ file of the project.operations.operational_types package.
This package is now required.

The results export functions of the project reserve provision modules
are now called directly from run_scenario.py instead of from
operations.py.

Modified the module prerequisites for serveral of the tests to reflect
the changes above. Added tests for the new modules.

Also with this commit: changed the summarize_results function in
power.py to reflect that the operations.csv results file has been
replaced by dispatch_all.csv, and that dispatch_all.csv includes the
timpoint, period, horizon, and horizon weight.