Add ramp constraints for disp cap commit
Ramp constraints for both up and down dispatch can now be optionally
implemented for projects of the 'dispatchable_capacity_commit'
operational type. The constraints are created if a ramp rate limit
(ramp_rate_up/down_frac_of_capacity) is specified in the projects.tab
input file. If not specified, the ramp rate param takes a default of 1,
which causes the constraints to be skipped.

The generators are allowed to turn on from 0 to Pmin and ramp up
above Pmin from one timepoint to the next (in the ramp up constraint)
and ramp down from above Pmin and turn off from Pmin to 0 from one
timepoint to the next (in the ramp down constraint).