Add min up/down time limits to disp cap commit
Track how much capacity was started up / shut down in each timepoint
relative to the previous timepoint, and ensure that at least this much
capacity  is online/offline respectively in the next N hours, where N
is the minimum up/down time. In other words, for each timepoint, we
will sum the capacity started up / shut down over previous timepoints
that fall within N hours of the current timepoint, and constraint
online / offline capacity to be greater than or equal to that the sum.