Allow storage losses to count in RPS
More accurately, allow storage losses to count against the RPS. This is
to avoid unrealistic storage behavior, e.g. storage behaving as load by
charging and discharging at the same time (and incurring losses) in
order to absorb RPS-eligible power that would otherwise be curtailed,
making it appear as if load is available that this power can be
delivered to. For this to work, storage projects must be flagged as
RPS-eligible (i.e. assigned an rps_zone).

By default all losses count against the RPS, but this can
be derated with the losses_factor_in_rps parameter. It is an open
question what this factor should be. Can we assume that all energy
absorbed by storage is RPS-eligible (storage is likely charging during
curtailment times) or is it more accurate to think of storage as
increasing load, therefore also increasing the RPS target?

As part of this commit, REC provision is split from power provision for
all operational types. This would allow for modeling unbundled RECs.