Initial implementation of a carbon cap
This is a simple implementation that tracks emissions from each project
assigned a 'carbon_cap_zone' and ensures that the sum of those
emissions over a period do not exceed a pre-specified carbon cap for
this carbon_cap_zone and period. Imports into a carbon cap zone are
currently not taken into account, so this simple approach is
appropriate only for modeling ALL carbonaceous projects as part of a
single carbon cap zone or at least as part of carbon cap zones,
each with its own carbon constraint. Otherwise, 'leakage' and
'reshuffling' of emissions would likely occur in the model.