Add 'technology' param
Add 'technology' param in the model -- loading data for it is optional,
as the param will have a default value of 'unspecified.' Having this
param simplifies aggregating results to technology later, as no join is
needed to an outside table. Results summaries are updated accordingly.
The dispatch and capacity modules will now export technology and
load_area with results automatically.

Also implemented updates to the capacity package, with the
capacity_type submodules now exporting their own results instead of
adding to a pandas dataframe in the main project.capacity.capacity
module. Merging the dataframes was very slow.