New database schema
Simplified the database schema and changed the way modules get inputs
from the database accordingly.

The script get_scenario_inputs.py is now in the root directory and is
responsible only for calling the scenario-appropriate modules, which
then get their own inputs, rather than running any queries itself.

A few column name changes in the input files result in minor changes to
the examples and tests.