Add ramp rate limit on reserves
Some reserve products may require that generators respond within a
certain amount of time, e.g 1 minute, 10 minutes, etc. The maximum
amount of reserves that a generator can provide is therefore limited
by how much it can ramp in that amount of time. The ramp rate limit
parameter is specified as a fraction of the generator's capacity and
defaults to 1 if not specified.