Add frequency response reserves
These mostly follow the standard reserve implemenation with a few
deviations due to the fact that a fraction of the reserve requirement
must be met by a subset of the generators that can provide it -- the
'partial' requirement. The partial requirement must be specified in
frequency_response_requirement.tab and the generators that can provide
must be flagged accordingy via the 'frequency_response_partial' column
in projects.tab.