Startup and shutdown costs per MW not unit
Changed the treatment of startup and shutdown costs to be per MW not
per unit (started up or shut down respectively).

The input numbers for startup and shutdown cost in the examples remain
the same, so the new treatment results in slight increases to the
objective functions of the examples.