Optionally constrain hydro ramps
Curtailable hydro and non-curtailable hydro ramps can now be
constrained if a 'ramp_up_when_on_rate' or a 'ramp_down_when_on_rate'
parameter is specified for the project in projects.tab.
This is optional -- the params will default to 1 if not specified and
the ramp constraints will be dropped. For curtailable hydro the total
energy in the current timepoint (i.e. power provided + power curtailed)
is used in the ramp constraints.