Add tuning for hydro ramps
Since the cost incurred by hydro over the course of a horizon is the
same regardless of exact dispatch pattern, cases may arise when the
project is ramped unnecessarily unless there's a cost on the ramp. This
is mostly an optics issue that can be addressed by putting a small
cost on hydro ramps. The cost defaults to 0 if not specified and the
associated constraints are dropped.

Tuning costs for hydro ramps and for imported carbon are now in the
database. The default 'import_carbon_tuning_cost' is now 0 to prevent
the potential to silently affect results without user knowledge.