Add endogenous dynamic ELCC
Projects can now contribute to the PRM either via a simple ELCC,
expressed as a fraction of their capacity, or via a dynamic ELCC
surface. The former is more appropriate for 'conventional' baseload
and dispatchable generators for which the main source of
'unavailability' during times of high system stress are unplanned
outages. Treatment via the ELCC surface is more appropriate for
'variable' generators whose ouptput during peak is uncertain even
without an outage. The idea behind the 'surface' is that the relative
penetration levels of these types of technologies will affect their
ELCC. For example, in a system with a daytime peak, solar initially
contributes a large fraction of its capacity to the PRM, as its highest
output conincides with system peak. However, as the solar penetration
level increases, the net load peak moves to the evening when solar is
unavailable, leading to declining marginal ELCC for solar. The relative
penetration levels of wind and solar affect their ELCCs. This is the
type of the dynamic that ELCC surface implemented here tries to capture.

The ELCC surface is implemented as a piecewise linear constraint. Each
project has its own coefficient for contributing to the surface in its
PRM zone (although coefficients are likely to be the same for projects
of the same technology, as surfaces of more than 2-3 dimensions would
be harder to create on the data development side) and the surface
intercept is defined by PRM zone.