Add threshold capacity costs
These are costs applied to new capacity only after a certain
threshold level of capacity is exceeded for each 'threshold group' of
projects. For example, new interconnection links -- with the associated
costs -- could not be needed for the first X MW installed, but can
become necessary for any additional capacity.

Such interconnection costs are currently the only type of 'threshold'
costs. The 'threshold groups' for these costs are assigned to projects
in the same table as load_zones (inputs_project_load_zones), as they
are dependent on how project geography is modeled.

'Threshold groups' can include a single project or multiple projects.