Get previous timepoint from order in TIMEPOINTS set (#31)
Instead of taking the timepoint's ID minus 1, the previous timepoint for each
timepoint is now derived from the timepoint's index in the TIMEPOINTS set, i.e.
by the taking the timepoint with index i-1 as the previous timepoint for the
timepoint with index i. For the first timepoint of a horizon, no previous
timepoint is set if the horizon has a linear boundary and the last timepoint of
the horizon is assigned as the previous timepoint if the horizon has a circular
boundary.