Fix minup/mindown constraint for capacity_commit
The current minimum up-time and minimum down-time constraints in the
capacity_commit operational module do not properly take into account
how varying timepoint durations will change what timepoints should
be included in the relevant_tmps when creating the constraint.

Rather than a for-loop based on the current timepoint's duration,
this update evaluates the elapsed time as you go back to previous
timepoints, until you hit the minimum up or down time.

This approach should also be used in the minimum up-time and minimum
down-time constraints for the binary_commit and continuous
commit operational types.

Note that this constraint is currently not tested in any of the
example scenarios, so this update does not affect the unittests.
We might want to think about adding a test case to the examples

Closes: #36