Add reserves-provision to always-on operational type
In theory always-on should be able to provide upward and/or downward
reserves if there is headroom and/or footroom. In the special case
that an always-on generator cannot provide reserves, you can simply
not provide a reserve zone in "projects.tab".

This update adds the reserve-provision variables to the minimum and
maximum generation constraints of always-on generators, ensuring these
operational types can provide reserves and are properly constrained.

Notes:
 - The new functionality is currently not tested in the unittests. We
   might want to add/modify a unittest to include an always-on generator
   with a reserve zone specified.
 - Separate PR will add reserve provision into ramp rates (issue #30)
 - Before this update, always-on generators could actually provide
   reserves if the user added a reserve zone to "projects.tab", but the
   variable would not be constrained. This bug is discussed in issue #39
   and will be addressed in separate PR

Closes: #41