Fix unconstrained reserve provision
Currently, operational types that are assumed not to provide
reserves have no constraints on their reserve provision in the
operational_type's module. However, if a user provides a reserve
sharing zone for any of these type of resources, the model will
create a decision variable for that type of reserve,
and the variable will be totally unconstrained.

This bug currently applies to the following operational types:
 - must_run
 - variable_no_curtailment
 - always_on (will get reserve provision so won't be an issue,
   see PR #43)

This commit solves the issue by adding a constraint for the must_run
and variable_no_curtailment operational types that sets the sum of
all headroom variables and the sum of all footroom variables to zero.

Another option would be to disallow inputting any reserve sharing zone
for certain type of generators (must_run, variable_no_curtailment)