Fix unconstrained reserve provision (#44)
Currently, operational types that are assumed not to provide
reserves have no constraints on their reserve provision in the
operational_type's module. However, if a user provides a reserve
sharing zone for any of these type of resources, the model will
create a decision variable for that type of reserve,
and the variable will be totally unconstrained.

This bug currently applies to the following operational types:
 - must_run
 - variable_no_curtailment
 - always_on (will get reserve provision so won't be an issue,
   see PR #43)

This PR solves the issue by adding a constraint for the must_run
and variable_no_curtailment operational types that sets the sum of
all headroom variables and the sum of all footroom variables to zero.

The PR also adds a warning when the user does provide a reserve BA
for these operational types erroneously, to make sure the user is aware 
of this issue and how it is addressed. 
In the future, we could address this issue into the input validation, and 
disallow inputting any reserve sharing zone for certain type of generators 
(must_run, variable_no_curtailment).

Closes: #39