Fix min up/down constraint for capacity_commit (#42)
This PR changes the minimum up/down time constraint for the capacity 
commit operational type such that it supports variable timepoint durations. 

The for-loop that determines the relevant timepoints to be included in the 
minimum up/down constraint, which was based on the current timepoint's 
duration, is replaced by a while loop that keeps track of the amount of hours 
you are removed from the current timepoint as you go back to previous
timepoints, and exits as soon as you reach the minimum up or down time. 

The PR also adds a common functions module for the operational type
modules and moves the "determine_relevant_timepoints" function for min 
up/down time constraints from the dispatchable_capacity_commit module to 
the common module that operational types can share. We also add a unittest
for "determine_relevant_timepoints" that explains and tests the behavior. 

Currently, this is implemented in the dispatchable_capacity_commit operational 
type module only, but it should be added to the binary_commit and 
continuous_commit operational types too (see PR #40 and issue #37). 

Note that this constraint is currently not tested in any of the example 
scenarios, so this update does not affect the examples unittests. We 
might want to think about adding a test case to the examples that includes
a generator with min up/down limits (although we are already testing the "determine_relevant_timepoints" common function). 

Closes: #36