Use common function in all min up/down constraints
This commit re-writes the minimum up and down constraints for
binary and continuous commit operational types so that they use
the common function "determine_relevant_timepoints" which is
also used in the capacity commit operational type. This
simplifies the constraint, makes it easier to unittest it,
and aligns it with the capacity commit module.