Add reserve provision to ramp rate constraints (#51)
When you are providing upward reserves in the current timepoint,
or downward reserves in the previous timepoint, you need to make
sure that your ramp up limits allow you to provide these reserves.
(and vice versa for downward/upward reserves and ramp down limits).
E.g. if you are ramping up your unit by 100 MW but you were also
providing 50 MW of downward reserves in the previous timepoint and
50 MW of upward reserves in the current timepoint, you need to make
sure your unit can ramp at least 200 MW (100 + 50 + 50).

This commit adjusts the ramp rate constraints to take into account
the reserve provisions in the current and previous timepoint. The
following operational types are affected:
 - always_on
 - dispatchable_capacity_commit
 - hydro_curtailable
 - hydro_noncurtailable

The binary and continuous commit operational types are being updated
as part of another PR (#50) and will have the proper adjustments in
their ramp rate constraints.

When projects provide reserves, the actual ramp will differ from the
difference between power setpoints. To avoid confusoin, we rename the
ramp_rule to power_delta_rule.

We also adjust the test examples for the ramp rates slightly. The
previous test examples resulted in unserved energy once we added
the adjustment for reserve provision to the ramp rates, which resulted
in a large change in the objective function. This update changes the
ramp rate limit inputs such that they still bind (and cause a slightly
more costly dispatch), but not so severely that they cause unserved
energy.