Fix incompatible types in test cases (#54)
Some of the test cases combined continuous commit operational types with
new build capacity types, which is not an intended use case, and creates
a non-linear problem. This commit changes the operational type of any
new build thermal generator to the capacity_commit type. This linearizes
the problem and removes the ipopt solver dependency of the test cases

Changes:
 - remove incompatible types: continuous commit and new build
 - remove ipopt solver settings in test examples
 - fix input error that set the min_stable_level at 0. Set it at 0.4
   instead. If set at zero, unit can avoid startup/shutdown costs
   by simply running at 0 MW instead of shutting down.
 - add heat rate inputs files for multistage prod cost with hydro
   test example (they were not being tracked by the VCS yet due to 
   new file structure)