change tolerance of infeasibility warnings
The default setting to warn users for infeasible constraints is 1E-6
which resulted in a lot of false-positive warnings about infeasible
constraints for the test examples. The difference between the
constraint's body and its bound would be e.g. 1.999E-6, which is not
an actual infeasibility but a rounding issue.

This commit changes the feasibility tolerance to 1E-3, which removes the
false positive warnings from the test examples.